+++ modular/signer/signer_client.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/base64"
 	"fmt"
-	"math/big"
 	"strings"
 	"sync"
 
@@ -300,6 +299,16 @@ func (client *GreenfieldChainSignClient) SealObjectEvm(ctx context.Context, scop
 		return "", ErrSignMsg
 	}
 
+	cosmosChainId, err := client.greenfieldClients[scope].GetChainID()
+	if err != nil {
+		return "", err
+	}
+
+	chainId, err := types.ParseChainID(cosmosChainId)
+	if err != nil {
+		return "", err
+	}
+
 	client.sealLock.Lock()
 	defer client.sealLock.Unlock()
 
@@ -315,7 +324,7 @@ func (client *GreenfieldChainSignClient) SealObjectEvm(ctx context.Context, scop
 	for i := 0; i < BroadcastTxRetry; i++ {
 		nonce = client.sealAccNonce
 
-		txOpts, err := CreateTxOpts(ctx, client.evmClient, client.privateKeys[SignSeal], big.NewInt(5151), client.gasInfo[Seal].GasLimit, nonce)
+		txOpts, err := CreateTxOpts(ctx, client.evmClient, client.privateKeys[SignSeal], chainId, client.gasInfo[Seal].GasLimit, nonce)
 		if err != nil {
 			log.CtxErrorw(ctx, "failed to create tx opts", "error", err)
 			return "", err